requirements.txt
streamlit
crewai
langchain_google_genai
python-dotenv